+++ app/views/admin/users/_user.html.erb
@@ -18,7 +18,7 @@
       <span id="row-<%= user.id %>" data-toggle="dropdown" aria-expanded="false" class='btn dropdown-toggle glyphicon glyphicon-align-justify'></span>
       <ul class="dropdown-menu dropdown-menu-right" role="menu">
         <li><%= link_to t(:discussions, default: 'Discussions'), admin_user_path(user.id), remote: true %></li>
-        <li><%= link_to t(:edit_user, default: 'Edit User'), admin_user_path(user.id, mode: 'edit'), remote: true %></li>
+        <li><%= link_to t(:edit_user, default: 'Edit User'), edit_admin_user_path(user.id, mode: 'edit'), remote: true %></li>
         <li><%#= link_to t(:open_new_discussion, default: 'Open Discussion'), admin_new_ticket_path, remote: true %></li>
       </ul>
     </div>
