+++ app/views/admin/users/_users.html.erb
@@ -3,7 +3,7 @@
   <col width="85%">
   <col width="10%">
 
-  <%= render partial: 'user', collection: @users if @users %>
+  <%= render partial: 'admin/users/user', collection: @users if @users %>
 </table>
 
 <div class="text-center">
