+++ app/views/admin/users/users.html.erb
@@ -3,6 +3,6 @@
 
 <div class="row">
   <div id="tickets">
-    <%= render 'users' %>
+    <%= render partial: 'admin/users/users' %>
   </div>
 </div>
