+++ app/views/layouts/_admin_header.html.erb
@@ -5,11 +5,11 @@
 
           <%= menu_group pull: :right do %>
             <% if @remote_search == true %>
-              <%= form_tag admin_search_path, method: :get, remote: true, class: 'navbar-form navbar-left' do %>
+              <%= form_tag admin_topic_search_path, method: :get, remote: true, class: 'navbar-form navbar-left' do %>
                 <%= text_field_tag('q', nil, placeholder: t(:search, default: 'Search'), value: params[:q], class: 'form-control full-width') %>
               <% end %>
             <% else %>
-              <%= form_tag admin_search_path, method: :get, class: 'navbar-form navbar-left' do %>
+              <%= form_tag admin_topic_search_path, method: :get, class: 'navbar-form navbar-left' do %>
                 <%= text_field_tag('q', nil, placeholder: t(:search, default: 'Search'), value: params[:q], class: 'form-control full-width') %>
               <% end %>
             <% end %>
